adding rating service

+++ src/main/java/layer/data/api/RatingData.java
@@ -1,9 +1,11 @@
 package layer.data.api;
 
+import java.math.BigDecimal;
+
 public interface RatingData {
 
   RatingRecord rate(Long idMovie);
 
-  void rate(Long idUser, Long idMovie, int value) throws DataException;
+  void rate(Long idUser, Long idMovie, BigDecimal value) throws DataException;
 
 }
